+++ user-center-backend-master/src/main/java/com/gasen/usercenterbackend/model/vo/PostDetail.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class PostDetail implements Serializable {
     // 帖子详情：appId、头像、标题、内容、点赞数、评论数、图片、创建时间、更新时间
     private Long id;
-    private Integer appId;
+    private Long appId;
     private String appName;
     private String avatar;
     private Integer grade;
