+++ user-center-backend-master/src/main/java/com/gasen/usercenterbackend/service/IItemsService.java
@@ -3,7 +3,7 @@ package com.gasen.usercenterbackend.service;
 import java.util.List;
 
 public interface IItemsService {
-    void addItem(Integer userId, Integer itemId);
+    void addItem(Long userId, Integer itemId);
 
-    List<Integer> getItems(int userId);
+    List<Integer> getItems(Long userId);
 }
